Got tick and basic boulder movement to work correcty
turns out timeline works in a different way than we thought it did
so i had to  create another one
i created  lists of each element , delete/add reference to object
created boulder tick , what happens here is it looks at the list of boulders and independently checks IF the boulder is allowed to fall and allows it to fall 1 cell/tick
redistributed the methods for code base
I also had to clear the arraylist for boulders at the initialization of the grid bcoz if i reset  and the ticks are on , it just keeps on adding boulders. just a PSA that sometimes u need to clear the memory(even tho it wasnt filling the grid from the list 🤦‍♂️ ).